requirements.txt
anyio==4.2.0
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
distlib==0.3.8
et-xmlfile==1.1.0
filelock==3.13.1
func-timeout==4.3.5
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==1.24.2
openpyxl==3.1.2
outcome==1.3.0.post0
pandas==1.5.3
platformdirs==4.1.0
protobuf==3.20.3
pycparser==2.21
PyMySQL==1.0.3
PySocks==1.7.1
python-dateutil==2.8.2
python-telegram-bot==20.8
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.9
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.2.0
virtualenv==20.25.0
wsproto==1.2.0
xlsx2csv==0.8.1
